% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREDICTbreastML.R
\name{PREDICTbreastML}
\alias{PREDICTbreastML}
\title{Machine-Learning Breast Cancer Survival Prediction (PREDICTbreastML)}
\usage{
PREDICTbreastML(ext_val_data = toyData, censor = 10)
}
\arguments{
\item{ext_val_data}{A data frame or data.table containing external validation
samples. If not supplied, the built-in `toyData` dataset is used.}

\item{censor}{censor time}
}
\value{
A numeric vector of predicted 10-year breast cancer–specific
  survival probabilities.
}
\description{
`PREDICTbreastML()` generates 10-year breast cancer–specific survival
predictions using a machine-learning survival model trained on the largest
and most diverse SEER cohort. The function automatically applies
ancestry-specific population reweighting, constructs a survival prediction
task, and returns the model-estimated probability of surviving to 10 years.
}
\examples{
data(toyData)
PREDICTbreastML(ext_val_data = toyData,censor = 10)

}
