% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyData.R
\docType{data}
\name{toyData}
\alias{toyData}
\title{toyData for testing PREDICTbreastML R package}
\format{
A R data frame with 10 rows and 10 variables, including survival time(fu), survival status(death_br_15), ancestry("Non-Hispanic Black", "Non-Hispanic White", "Hispanic", "Non-Hispanic Asian"),age at diagnosis, ER status, PR status, tumor grade, positive nodes, tumor size and mode of detection
\describe{
  \item{clinical data}{Numeric clinical matrix with each
  rows as a different individual and each column as a clinical feature.}}
}
\source{
<https://www.github.com/pengpclab/PREDICTbreastML>
}
\usage{
data(toyData)
}
\description{
A simulated data set with clinical information for 10 breast cancer cases
}
\examples{
data(toyData)
}
\keyword{datasets}
